# pkg/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"net/http"
)

// Media defines model for Media.
type Media struct {
	// Name Name of the media item
	Name string `json:"name"`

	// Tags List of tag names associated with the media
	Tags []string `json:"tags"`

	// Url URL to access the media file
	Url string `json:"url"`
}

// Tag Name of the tag
type Tag = string

// UploadRequest defines model for UploadRequest.
type UploadRequest struct {
	// Method HTTP method to be used (usually PUT or POST)
	Method string `json:"method"`

	// SignedHeader HTTP headers required for the request
	SignedHeader http.Header `json:"signedHeader"`

	// Url The pre-signed URL to upload the media
	Url string `json:"url"`
}

// GetMediaParams defines parameters for GetMedia.
type GetMediaParams struct {
	// Tag Tag to search for media items
	Tag string `form:"tag" json:"tag"`
}

// PostMediaJSONBody defines parameters for PostMedia.
type PostMediaJSONBody struct {
	// Name Name of the media item
	Name string `json:"name"`

	// Tags List of tags associated with the media
	Tags []string `json:"tags"`
}

// PostTagsJSONBody defines parameters for PostTags.
type PostTagsJSONBody struct {
	// Name Name of the tag
	Name string `json:"name"`
}

// PostMediaJSONRequestBody defines body for PostMedia for application/json ContentType.
type PostMediaJSONRequestBody PostMediaJSONBody

// PostTagsJSONRequestBody defines body for PostTags for application/json ContentType.
type PostTagsJSONRequestBody PostTagsJSONBody